update bank accounts components
