+++ src/components/bank-accounts/list/destroy.jsx
@@ -1,0 +1,51 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+import { withRouter } from 'react-router-dom';
+import { Button } from 'react-bootstrap';
+
+import { addFlashMessage } from 'actions/flash-messages.js';
+import { deleteBankAccount } from 'actions/bank-accounts.js';
+import { getCurrentOrganizationId } from 'selectors/organizations.js';
+
+class DestroyBankAccount extends React.Component {
+  constructor(props) {
+    super(props);
+    this.handleDeleteBankAccountClick = this.handleDeleteBankAccountClick.bind(this);
+  }
+
+  handleDeleteBankAccountClick() {
+    const { orgId, bankAccount, deleteBankAccount } = this.props;
+    deleteBankAccount(orgId, bankAccount.id).then(bankAccount => {
+      this.props.addFlashMessage('Bank account ' + bankAccount.name + ' successfully deleted.');
+      this.props.history.push('/bank_accounts');
+    }).catch(error => {
+      this.props.addFlashMessage(`Unable to delete bank account: ${error.message}`, { type: 'danger' });
+    });
+  }
+
+  render() {
+    return (
+      <Button bsStyle="danger" onClick={ this.handleDeleteBankAccountClick() }>Delete</Button>
+    );
+  }
+}
+
+DestroyBankAccount.propTypes = {
+  orgId:             PropTypes.number.isRequired,
+  deleteBankAccount: PropTypes.func.isRequired,
+  bankAccount:       PropTypes.object.isRequired,
+  addFlashMessage:   PropTypes.func.isRequired,
+  history:           PropTypes.object.isRequired,
+};
+
+const select = (state) => ({
+  orgId: getCurrentOrganizationId(state),
+});
+
+const dispatcher = (dispatch) => ({
+  deleteBankAccount: (organizationId, bankAccountId) => new Promise((res, rej) => dispatch(deleteBankAccount(organizationId, bankAccountId, res, rej))),
+  addFlashMessage:   (message, type = null) => dispatch(addFlashMessage(message, type)),
+});
+
+export default withRouter(connect(select, dispatcher)(DestroyBankAccount));
