+++ src/components/bank-accounts/list/table-body.jsx
@@ -1,0 +1,10 @@
+import * as React from 'react';
+import Row from './table-row.jsx';
+
+const BankAccountsTableBody = ({ bankAccounts }) => (
+  <tbody>
+    { bankAccounts.map((ba) => <Row bankAccount={ ba } key={ ba.id } />) }
+  </tbody>
+);
+
+export default BankAccountsTableBody;
